feat: move queries to a selet package
