+++ jnosql-communication/jnosql-communication-query/src/main/java/org/eclipse/jnosql/communication/query/data/DefaultQueryValue.java
@@ -16,7 +16,8 @@ import org.eclipse.jnosql.communication.query.ParamQueryValue;
 
 /**
  * The default implementation of {@link ParamQueryValue}
- */ record DefaultQueryValue(String value) implements ParamQueryValue {
+ */
+public record DefaultQueryValue(String value) implements ParamQueryValue {
 
 
     @Override
